--- Case/__init__.py
@@ -1,2 +1,0 @@
-"""package containg tool to construct Cases, display and select its range of Intervals"""
-import case.intervals
