+++ kg/Case/__init__.py
@@ -1,0 +1,3 @@
+"""package containg tool to construct Cases, display and select its range of Intervals"""
+import Case.intervals
+import Case.case
