+++ kg/__init__.py
@@ -8,6 +8,5 @@ The kg module.
 import kg.time_signal
 import kg.measurement_values
 import kg.mpl_moving_bar
-import kg.intervals
-import kg.case
+import kg.Case
