+++ kg/measurement_values.py
@@ -6,94 +6,10 @@ import copy
 import collections
 import xlrd
 import csv,json
-import datetime
 import pathlib
 import warnings
-import datetime
-
-def read_MBBM_tables(mesPath, save = False):
-    '''
-    load values of selected variables and mic form MBBM files
-    '''
-    mesPath = pathlib.Path(mesPath)
-    with mesPath.joinpath('measurement_config.json').open('r+') as config:
-        CONFIG = json.load(config)
-    TABLES = CONFIG['tables']
-    MICVALUES = CONFIG['micValues']
-    IDVALUES = CONFIG['idValues']
-    Mic = CONFIG['microphones']
-    tablesPath = mesPath.joinpath('measurement_values')
-    print('----------\n')
-    print('Read MBBM exel tables \n----------\n\nMIC VALUES \n-----------')
-    for k,v in MICVALUES.items():
-        print(k,', ', end='')
-        table = TABLES[v['table']]
-        wb = xlrd.open_workbook(str(tablesPath.joinpath(table['fileName'])))
-        #if Spectrum
-        if k == 'LAf':
-            S_mic = []
-            data = {}
-            for mic in Mic:
-                sN = v['sheet'].replace('_Miki','_Mik'+ str(mic))
-                ws = wb.sheet_by_name(sN)
-                for row in range(table['skip']+1,ws.nrows):
-                    S_mic = [ws.cell_value(row,i) for i in range(1,25)]
-                    try:
-                        dictID = data[ws.cell_value(row,0)]
-                    except KeyError:
-                        data[ws.cell_value(row,0)] = {}
-                        dictID = data[ws.cell_value(row,0)]
-                    dictID[str(mic)] = S_mic
-            v['values'] = data
-        #if mic single valued value
-        else:
-            #ID column
-            ws = wb.sheet_by_name(v['sheet'])
-            #ws starting row number of  
-            wsRow0 = table['skip']
-            # col names Table in excel ws
-            colNames = ws.row_values(wsRow0)
-            #index of selectd columns
-            cN = [(v['colName']).replace('_i','_'+ str(mic)) for mic in Mic]
-            selectCol = [colNames.index(i) for i in cN] 
-            data = {}
-            for row in range(wsRow0+1,ws.nrows):
-                mic_val = {}
-                for i,mic in zip(selectCol,Mic):
-                    mic_val[str(mic)] = ws.cell_value(row,i) 
-                data[ws.cell_value(row,0)] = mic_val
-            v['values'] = data
-    #if IDVALUES value
-    print('\n\nIDVALUES \n-----------')        
-    for k,v in IDVALUES.items():
-        print(k,', ', end='')
-        table = TABLES[v['table']]
-        wb = xlrd.open_workbook(str(tablesPath.joinpath(table['fileName'])))
-        ws = wb.sheet_by_name(v['sheet'])
-        # col names Table
-        colNames = ws.row_values(table['skip'])
-        selectCol = colNames.index(v['colName']) 
-        data = {}
-        for row in range(table['skip']+1, ws.nrows):
-            data[ws.cell_value(row,0)] = ws.cell_value(row,selectCol)
-        v['values'] = data
-    print('\n\nFinish read values\n----------\n')
-    # return dict
-    MBBM_data = collections.OrderedDict()
-    MBBM_data['location'] = CONFIG['location']
-    MBBM_data['measurement'] = CONFIG['measurement']
-    MBBM_data['tables'] = TABLES
-    MBBM_data['micValues'] = MICVALUES
-    MBBM_data['idValues'] = IDVALUES
-    MBBM_data['mic'] = Mic
-    if save:
-        dataPath = mesPath.joinpath('measurement_values/MBBM_mes_values.json')
-        with dataPath.open('w+') as data:
-            json.dump(MBBM_data, data)
-        print('Data saved to ' + mesPath.as_posix())
-    return(MBBM_data)
-        
-
+import time
+## Class measuredValues
 class measuredValues():
     def __init__(self,mesPath, location, measurement, tables, micValues, idValues , mic):
         '''
@@ -129,6 +45,18 @@ class measuredValues():
         else:
             Index = self.micValues['LAEQ']['values'].index.tolist()
         return(Index)
+    
+    def getVariable(self,ID, mic, Variable):
+        """
+        fetch the variable given in Variables from self for the ID given in ID and the microphone given in mic
+        """
+        retvalue=0
+        if Variable in self.micValues.keys():
+            retvalue=self.micValues[Variable].get('values').get(ID,{}).get(str(mic),{})
+        else:
+            print('Variable '+vars+' not ok. Please use `get_variables_values(self,ID,mic,variables)`, method of class measuredValue')
+            retvalue=None
+        return retvalue
         
     def get_variables_values(self,ID, mic, variables):
         """
@@ -220,7 +148,8 @@ class measuredValues():
     @classmethod    
     def from_MBBM(cls, mesPath):
         return(cls(mesPath = mesPath,**read_MBBM_tables(mesPath, save = False)))
-        
+
+##functions
 def serialize(data):
     if isinstance(data, list):
         return [serialize(val) for val in data]
@@ -233,14 +162,101 @@ def serialize(data):
     else:
         return(data)
 
-if __name__ == "__main__":
-    
-    AA = measuredValues('C:\lucmiaz\KG_dev_branch\KG\Measurements_example\MBBMZugExample')
+def read_MBBM_tables(mesPath, save = False):
+    '''
+    load variables and mic from MBBM files in path=mesPath
+    '''
+    mesPath = pathlib.Path(mesPath)
+    with mesPath.joinpath('measurement_config.json').open('r+') as config:
+        CONFIG = json.load(config)
+    TABLES = CONFIG['tables']
+    MICVALUES = CONFIG['micValues']
+    IDVALUES = CONFIG['idValues']
+    Mic = CONFIG['microphones']
+    tablesPath = mesPath.joinpath('measurement_values')
+    print('----------\n')
+    print('Read MBBM exel tables \n----------\n\nMIC VALUES \n-----------')
+    for k,v in MICVALUES.items():
+        print(k,', ', end='')
+        table = TABLES[v['table']]
+        wb = xlrd.open_workbook(str(tablesPath.joinpath(table['fileName'])))
+        #if Spectrum
+        if k == 'LAf':
+            S_mic = []
+            data = {}
+            for mic in Mic:
+                sN = v['sheet'].replace('_Miki','_Mik'+ str(mic))
+                ws = wb.sheet_by_name(sN)
+                for row in range(table['skip']+1,ws.nrows):
+                    S_mic = [ws.cell_value(row,i) for i in range(1,25)]
+                    try:
+                        dictID = data[ws.cell_value(row,0)]
+                    except KeyError:
+                        data[ws.cell_value(row,0)] = {}
+                        dictID = data[ws.cell_value(row,0)]
+                    dictID[str(mic)] = S_mic
+            v['values'] = data
+        #if mic single valued value
+        else:
+            #ID column
+            ws = wb.sheet_by_name(v['sheet'])
+            #ws starting row number of  
+            wsRow0 = table['skip']
+            # col names Table in excel ws
+            colNames = ws.row_values(wsRow0)
+            #index of selectd columns
+            cN = [(v['colName']).replace('_i','_'+ str(mic)) for mic in Mic]
+            selectCol = [colNames.index(i) for i in cN] 
+            data = {}
+            for row in range(wsRow0+1,ws.nrows):
+                mic_val = {}
+                for i,mic in zip(selectCol,Mic):
+                    mic_val[str(mic)] = ws.cell_value(row,i) 
+                data[ws.cell_value(row,0)] = mic_val
+            v['values'] = data
+    #if IDVALUES value
+    print('\n\nIDVALUES \n-----------')        
+    for k,v in IDVALUES.items():
+        print(k,', ', end='')
+        table = TABLES[v['table']]
+        wb = xlrd.open_workbook(str(tablesPath.joinpath(table['fileName'])))
+        ws = wb.sheet_by_name(v['sheet'])
+        # col names Table
+        colNames = ws.row_values(table['skip'])
+        selectCol = colNames.index(v['colName']) 
+        data = {}
+        for row in range(table['skip']+1, ws.nrows):
+            data[ws.cell_value(row,0)] = ws.cell_value(row,selectCol)
+        v['values'] = data
+    print('\n\nFinish read values\n----------\n')
+    # return dict
+    MBBM_data = collections.OrderedDict()
+    MBBM_data['location'] = CONFIG['location']
+    MBBM_data['measurement'] = CONFIG['measurement']
+    MBBM_data['tables'] = TABLES
+    MBBM_data['micValues'] = MICVALUES
+    MBBM_data['idValues'] = IDVALUES
+    MBBM_data['mic'] = Mic
+    if save:
+        dataPath = mesPath.joinpath('measurement_values/MBBM_mes_values.json')
+        with dataPath.open('w+') as data:
+            json.dump(MBBM_data, data)
+        print('Data saved to ' + mesPath.as_posix())
+    return(MBBM_data)
+##tests
+if __name__=="__main__":
+    for i in range(0,3):
+        start_time=time.time()
+        AA = measuredValues.from_json('C:\lucmiaz\KG_dev_branch\KG\Measurements_example\MBBMZugExample')
+        print("---%s seconds---" % (time.time()-start_time))
     AA.list_variables()
-    AA.read_variables_values()
+    
+    #AA.read_variables_values()
     ##getexample
-    s=AA.get_variables_values(ID='m_0100',mic= [1,2], variables=['v2','v1','direction','Ta', 'Te', 'Tp_a', 'Tp_e'])
-    print(s)
+    print("get_variables_values")
+    s=AA.get_variables_values(['m_0100','m_0101','m_0102','m_0103','m_0104','m_0105','m_0106','m_0107','m_0108','m_0109','m_0110','m_0111','m_0112','m_0113','m_0114','m_0115','m_0116','m_0117','m_0118','m_0119','m_0120','m_0121','m_0122','m_0123','m_0124','m_0125','m_0126','m_0127','m_0128','m_0129','m_0130','m_0131','m_0132','m_0133','m_0134','m_0135','m_3001','m_3011','m_3015','m_3021''k2000'],[1,2,3], ['v2','v1','direction','Tb', 'Te', 'Tp_b', 'Tp_e'])
+    l=AA.getVariable('m_0100',1,'Tb')
+    print(l)
     ##
